The YAML Update
Lots of changes, got rid of the config.txt file to use config.yaml, the
RakeFile and createApp2.rb files have been altered to reflect that
change, along with adding the implementation of Extras to allow special
elements in the app/dashboard. The overlay Extra has been implemented
and is available.